Refactored the new methods into a separate class and added documentation
